+++ model/PopupPosition.ts
@@ -1,0 +1,6 @@
+export enum PopupPosition {
+  BelowText = "BelowText",
+  AboveText = "AboveText",
+  LeftOfText = "LeftOfText",
+  RightOfText = "RightOfText",
+}