+++ pages/settings.tsx
@@ -22,6 +22,7 @@ import {isValidWebSocketUrl} from "../util/Url"
 import RestoreDefaultValueButton from "../components/RestoreDefaultValueButton"
 import {Api} from "../util/Api"
 import JpdbRulesEditModal from "../components/JpdbRulesEditModal"
+import PopupPositionSelect from "../components/PopupPositionSelect"
 
 interface Props {
   appSettings: AppSettings,
@@ -32,10 +33,12 @@ export default function Settings({appSettings, defaultAppSettings}: Props) {
   const toast = useToast()
 
   const [readingTimerEnabled, setReadingTimerEnabled] = useState(appSettings.readingTimerEnabled)
+  const [textHookerWebSocketUrl, setTextHookerWebSocketUrl] = useState(appSettings.textHookerWebSocketUrl)
   const [jpdbApiKey, setJpdbApiKey] = useState(appSettings.jpdbApiKey)
   const [jpdbMiningDeckId, setJpdbMiningDeckId] = useState(appSettings.jpdbMiningDeckId)
   const [jpdbRules, setJpdbRules] = useState(appSettings.jpdbRules)
-  const [textHookerWebSocketUrl, setTextHookerWebSocketUrl] = useState(appSettings.textHookerWebSocketUrl)
+  const [jpdbHorizontalTextPopupPosition, setJpdbHorizontalTextPopupPosition] = useState(appSettings.jpdbHorizontalTextPopupPosition)
+  const [jpdbVerticalTextPopupPosition, setJpdbVerticalTextPopupPosition] = useState(appSettings.jpdbVerticalTextPopupPosition)
 
   const [jpdbRulesEditPending, setJpdbRulesEditPending] = useState(false)
 
@@ -50,10 +53,12 @@ export default function Settings({appSettings, defaultAppSettings}: Props) {
     }
     await Api.updateAppSettings({
       readingTimerEnabled: readingTimerEnabled,
+      textHookerWebSocketUrl: textHookerWebSocketUrl,
       jpdbApiKey: jpdbApiKey,
       jpdbMiningDeckId: jpdbMiningDeckId,
       jpdbRules: jpdbRules,
-      textHookerWebSocketUrl: textHookerWebSocketUrl,
+      jpdbHorizontalTextPopupPosition: jpdbHorizontalTextPopupPosition,
+      jpdbVerticalTextPopupPosition: jpdbVerticalTextPopupPosition,
     })
     toast({
       description: "Settings saved",
@@ -81,6 +86,7 @@ export default function Settings({appSettings, defaultAppSettings}: Props) {
           <Container maxW="xl">
             <VStack alignItems="center" spacing="8">
               <Text fontSize="2xl">Settings</Text>
+
               <Text fontSize="xl">General</Text>
               <VStack alignSelf="start">
                 <Checkbox isChecked={readingTimerEnabled}
@@ -88,29 +94,8 @@ export default function Settings({appSettings, defaultAppSettings}: Props) {
                   Enable reading timer
                 </Checkbox>
               </VStack>
-              <Text fontSize="xl">Integrations</Text>
-              <FormControl>
-                <FormLabel>JPDB API key</FormLabel>
-                <HStack>
-                  <Input type="password" value={jpdbApiKey} onChange={event => setJpdbApiKey(event.target.value)}/>
-                  <RestoreDefaultValueButton onClick={() => setJpdbApiKey(defaultAppSettings.jpdbApiKey)}/>
-                </HStack>
-                <FormHelperText>JPDB API key used for text parsing and words highlighting.</FormHelperText>
-              </FormControl>
-              <FormControl>
-                <FormLabel>JPDB mining deck ID</FormLabel>
-                <HStack>
-                  <Input type="number" value={jpdbMiningDeckId}
-                         onChange={event => setJpdbMiningDeckId(parseInt(event.target.value))}/>
-                  <RestoreDefaultValueButton onClick={() => setJpdbMiningDeckId(defaultAppSettings.jpdbMiningDeckId)}/>
-                </HStack>
-                <FormHelperText>Mined words will be added to JPDB deck with this ID number.</FormHelperText>
-              </FormControl>
-              <FormControl>
-                <FormLabel>JPDB rules</FormLabel>
-                <Button onClick={() => setJpdbRulesEditPending(true)}>Edit</Button>
-                <FormHelperText>Configure colors and which words should be highlighted.</FormHelperText>
-              </FormControl>
+
+              <Text fontSize="xl">Text hooker</Text>
               <FormControl>
                 <FormLabel>Text hooker WebSocket URL</FormLabel>
                 <HStack>
@@ -121,7 +106,52 @@ export default function Settings({appSettings, defaultAppSettings}: Props) {
                 </HStack>
                 <FormHelperText>WebSocket URL used for the text hooker page.</FormHelperText>
               </FormControl>
-              <Box pt={4}>
+
+              <Text fontSize="xl">JPDB integration</Text>
+              <FormControl>
+                <FormLabel>API key</FormLabel>
+                <HStack>
+                  <Input type="password" value={jpdbApiKey} onChange={event => setJpdbApiKey(event.target.value)}/>
+                  <RestoreDefaultValueButton onClick={() => setJpdbApiKey(defaultAppSettings.jpdbApiKey)}/>
+                </HStack>
+                <FormHelperText>JPDB API key used for text parsing and words highlighting.</FormHelperText>
+              </FormControl>
+              <FormControl>
+                <FormLabel>Mining deck ID</FormLabel>
+                <HStack>
+                  <Input type="number" value={jpdbMiningDeckId}
+                         onChange={event => setJpdbMiningDeckId(parseInt(event.target.value))}/>
+                  <RestoreDefaultValueButton onClick={() => setJpdbMiningDeckId(defaultAppSettings.jpdbMiningDeckId)}/>
+                </HStack>
+                <FormHelperText>Mined words will be added to this JPDB deck.</FormHelperText>
+              </FormControl>
+              <FormControl>
+                <FormLabel>Horizontal text popup position</FormLabel>
+                <HStack>
+                  <PopupPositionSelect
+                    value={jpdbHorizontalTextPopupPosition}
+                    onChange={setJpdbHorizontalTextPopupPosition}/>
+                  <RestoreDefaultValueButton
+                    onClick={() => setJpdbHorizontalTextPopupPosition(defaultAppSettings.jpdbHorizontalTextPopupPosition)}/>
+                </HStack>
+              </FormControl>
+              <FormControl>
+                <FormLabel>Vertical text popup position</FormLabel>
+                <HStack>
+                  <PopupPositionSelect
+                    value={jpdbVerticalTextPopupPosition}
+                    onChange={setJpdbVerticalTextPopupPosition}/>
+                  <RestoreDefaultValueButton
+                    onClick={() => setJpdbVerticalTextPopupPosition(defaultAppSettings.jpdbVerticalTextPopupPosition)}/>
+                </HStack>
+              </FormControl>
+              <FormControl>
+                <FormLabel>Rules</FormLabel>
+                <Button onClick={() => setJpdbRulesEditPending(true)}>Edit</Button>
+                <FormHelperText>Configure colors and which words should be highlighted.</FormHelperText>
+              </FormControl>
+
+              <Box pt={8}>
                 <Button variant="solid" colorScheme="blue" onClick={saveSettings}>
                   Save settings
                 </Button>